Chapter 7: Evaluating Expressions (#4)
This implements [chapter 7](http://craftinginterpreters.com/evaluating-expressions.html).
We introduce a simple evaluation of some basic operators.

The biggest difference to the original Java version is that we are not setting `hadRuntimeError`
but return a `Result<String, Error>` from the `interpret` function.

The Java version also casts the literal to `String` or `Double` and handles exceptions. In Rust
we can simply pattern match and return a `Error::Runtime` when `+` is called for other types.