Chapter 10: Functions (#7)
This patch implements the [chapter on functions](http://craftinginterpreters.com/functions.html).

Instead of using an interface `LoxCallable` and and anonymous instance for native functions
I opted for an enum with the variants `Native` and `User`. Native functions are defined with a
function pointer and user functions are implemented the same way `LoxFunction`s are
implemented in the book.

The biggest difference is that we use `Err` and `Result` to jump back out of a call stack as
we don't have `Exception`s.