Chapter 11: Resolving and Binding (#8)
This implements [chapter 11](http://craftinginterpreters.com/resolving-and-binding.html) to resolve variables.

The resolver is probably a little odd for Rust and has two sources for bugs. For one the hash
map of the scopes uses the token instead of the expression for locals. If two tokens are on the
the same line this might be an issue. For another the `had_error` is still not implemented.