Chapter 12: Classes (#9)
This implements [chapter 12](http://craftinginterpreters.com/classes.html) to support classes.

In this implementation classes do not just extend `LoxFunction`. That's why the call logic is duplicated. Mutable state such as the environments and class instances are passed around in an `Object` that holds an `Rc<RefCell<...>>` to the actual state. It is not easy to track but it works. When an `Object` is cloned the reference is cloned. This mirrors somewhat how the
Java implementation works but is not really "rusty".

There are a lot of clone calls but it should not cost too much overhead since we mostly clone 
references and vectors.