Chapter 18: Types of Values (#15)
This implements chapter 18.

The biggest difference to the C version in the value type. I used functions instead of macros
to check the value type and casting. However, the C-like union is straight from the book. It is
a learning exercise. Rust's enums are tagged unions. A production ready implementation
should use enums.